third_party/blink/web_tests/fast/dom/timer-throttling-hidden-page-expected.txt

Bug 98474: Throttle DOM timers on hidden pages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS totalTime is >= 3 * requestedInterval + 2 * timerAlignmentInterval
PASS 1000 * 60 * 60 * 24 is >= totalTime
PASS successfullyParsed is true

TEST COMPLETE


This test measures the time taken to fire a 100ms DOM Timer when the page visibility is set to "visible", "hidden", and then back to "visible". Due to page timer throttling, the timer should fire close to 1s when page is hidden. And it should fire close to 100ms, when the page is visible.
